add authenticate and pagination

+++ controllers/contactsControllers.js
@@ -2,11 +2,22 @@ import Contact from "../models/contacts.js";
 
 import HttpError from "../helpers/HttpError.js";
 
-
-
 export const getAllContacts = async (req, res, next) => {
   try {
-    const result = await Contact.find();
+    const { _id: owner } = req.user;
+    const { page = 1, limit = 20, favorite } = req.query;
+    const skip = (page - 1) * limit;
+
+    let result;
+    if (favorite) {
+      result = await Contact.find({ owner, favorite }, "-_", {
+        skip,
+        limit,
+      });
+    } else {
+      result = await Contact.find({ owner }, "-_", { skip, limit });
+    }
+
     res.status(200).json(result);
   } catch (error) {
     next(error);
@@ -16,11 +27,14 @@ export const getAllContacts = async (req, res, next) => {
 export const getOneContact = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const result = await Contact.findById(id);
-    if (!result) {
+    const contact = await Contact.findById(id);
+    if (!contact) {
       throw HttpError(404);
     }
-    res.status(200).json(result);
+    if (!contact.owner || contact.owner.toString() !== req.user.id) {
+      throw HttpError(404);
+    }
+    res.status(200).json(contact);
   } catch (error) {
     next(error);
   }
@@ -29,10 +43,14 @@ export const getOneContact = async (req, res, next) => {
 export const deleteContact = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const result = await Contact.findByIdAndDelete(id);
-    if (!result) {
+    const contact = await Contact.findById(id);
+    if (!contact) {
       throw HttpError(404);
     }
+    if (!contact.owner || contact.owner.toString() !== req.user.id) {
+      throw HttpError(404);
+    }
+    const result = await Contact.findByIdAndDelete(id);
     res.status(200).json(result);
   } catch (error) {
     next(error);
@@ -41,7 +59,8 @@ export const deleteContact = async (req, res, next) => {
 
 export const createContact = async (req, res, next) => {
   try {
-    const result = await Contact.create(req.body);
+    const { _id: owner } = req.user;
+    const result = await Contact.create({ ...req.body, owner });
     res.status(201).json(result);
   } catch (error) {
     next(error);
@@ -51,10 +70,18 @@ export const createContact = async (req, res, next) => {
 export const updateContact = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
-    if (!result) {
+    const contact = await Contact.findById(id);
+    if (!contact) {
       throw HttpError(404);
     }
+    if (!contact.owner || contact.owner.toString() !== req.user.id) {
+      throw HttpError(404);
+    }
+
+    const result = await Contact.findByIdAndUpdate(id, req.body, {
+      new: true,
+    });
+
     res.status(200).json(result);
   } catch (error) {
     next(error);
@@ -64,10 +91,14 @@ export const updateContact = async (req, res, next) => {
 export const updateStatusContact = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
-    if (!result) {
+    const contact = await Contact.findById(id);
+    if (!contact) {
       throw HttpError(404);
     }
+    if (!contact.owner || contact.owner.toString() !== req.user.id) {
+      throw HttpError(404);
+    }
+    const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
     res.status(200).json(result);
   } catch (error) {
     next(error);
